[Quant][fx] Enable FX static quantization for LSTM
Summary: Static quantization for LSTM was previously only
supported in eager mode. This is because we had to go through
the custom module flow(torch.ao.nn.quantizable.LSTM). However,
this flow returned multiple outputs, and this is currently not
supported in FX graph mode quantization.

In this commit, we work around this limitation by making two
important modifications to this flow:

(1) First, the output of the LSTM node is in the form (result,
(hidden0, hidden1)), and each of the three internal nodes was
already observed, so there is actually no need to observe the
output again. This commit removes the extra output observer that
tried to observe the whole tuple, which failed previously.

(2) Second, the inputs of the LSTM node must be quantized, and
the outputs must be dequantized. This was not handled correctly
previously because FX graph mode quantization did not had special
logic for tuples. For the output in particular, we had to insert
ops to manually split the tuple, insert dequantize nodes, and
recombine them into the original format.

Note that the changes here are intended only as temporary hacks
to enable this flow. In the future, we should add better support
for dtype inference and tuple handling in FX graph mode quantization.

Test Plan:
python test/test_quantization.py TestQuantizeFx.test_static_lstm

Reviewers: jerryzh168, vkuzo

Subscribers: jerryzh168, vkuzo

ghstack-source-id: 79ba2fad7bbc8a248b4acd86ade4ecbc36da38f9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85068